[YS-151] refactor: 컴포넌트 스타일 분리
